refactor(BO-541): Extract IdempotentPersistentSubscriptionListener.kt to separate class.
This should improve testability since we can test method handling in isolation
