+++ build.gradle
@@ -1,6 +1,6 @@
 plugins {
-	id 'org.jetbrains.kotlin.jvm' version '1.3.21'
-	id 'org.jetbrains.kotlin.plugin.spring' version '1.3.21'
+	id 'org.jetbrains.kotlin.jvm' version '1.3.70'
+	id 'org.jetbrains.kotlin.plugin.spring' version '1.3.70'
 }
 
 apply plugin: 'maven-publish'
@@ -30,6 +30,7 @@ dependencies {
 	}
 	testImplementation 'org.junit.jupiter:junit-jupiter-api:5.3.1'
 	testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.3.1'
+	testImplementation 'io.mockk:mockk:1.9.3'
 }
 
 compileKotlin {
